# ARROW-4735: [Go] Optimize CSV writer CPU/Mem performances