This CL replaces all the naive calls to fmt.Sprintf with
strconv.FormatXYZ.

```    
$> benchstat old.txt new.txt
name     old time/op    new time/op    delta
Write-4    2.21ms ±12%    1.15ms ± 1%  -47.78%  (p=0.000 n=30+30)

name     old alloc/op   new alloc/op   delta
Write-4     335kB ± 0%     308kB ± 0%   -8.21%  (p=0.000 n=30+29)

name     old allocs/op  new allocs/op  delta
Write-4     21.8k ± 0%     11.6k ± 0%  -46.96%  (p=0.000 n=30+30)
```