@anson627 2019-03-12
LGTM, is Sprintf slow becaues of using generics?