@sbinet 2019-03-12
`fmt.Sprintf` is slower than `strconv.FormatXYZ` because it needs to:
- be generic (but there are no generics yet in Go)
- apply type introspection on the value(s) being passed as arguments to decide how to format them
- parse the format to know how the values have to be formatted as strings.

by using directly `strconv.FormatXYZ` directly, we can shortcut a bunch of CPU and allocations.